Separate OpenApi documentation to new module
